requirements.txt
Django == 5.1.4
web3 == 7.6.0
python-decouple == 3.8
